Upgrade to newer reason-react version
I had to upgrade the reason-react version to 0.2.4 to follow your tutorial. The correct version is documented in the tutorial, but it took me a while to find why it doesn't compile. It would be great if the starter project came directly with 0.2.4.

And thanks a lot for the tutorial!